requirements.txt
# --- core ---
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

lightning==2.6.0
torchmetrics==1.8.2

numpy==2.0.1
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.15.3

matplotlib==3.10.8
PyYAML==6.0.3
tqdm==4.67.1
pillow==11.1.0

# --- visuals / retrieval (we use them in this repo) ---
umap-learn==0.5.9.post2
faiss-cpu==1.13.2
rich==14.2.0
